Improved Linux LLVM/libclang detection
The previous path detection logic of pybind11_mkdoc uses mismatched LLVM
and libclang paths e.g. when LLVM 9 & 11 is installed but the libc++
version is stuck at v9.

This commit causes the two to be consistent and generally simplifies the
code that has grown somewhat organically. The ability to override paths
is retained.