Add support for the libclang package
A newer package provigind the python libclang bindings is available, called
"libclang". Unlike "clang", this package ships the libclang library, removing
the dependency from the system's one.

This commit adds support for it, using the libclang's builtin library if found
and if LIBCLANG_PATH and LLVM_DIR_PATH are not defined.

As libclang does not ship the clang standard library headers, some system
headers may not be found (like stddef.h), but the documentation generation
completes successfully anyways.